Revert ControlNetInvocation changes.
